Emit an error log line if client writer fails writing (SSL_write()).
Also, don't try to write zero number of octets because OpenSSL might
not like that.

I would like to close the connection too but would have to look into
the UDP and DTLS cases more before that can be done.

This is for figuring out more about how to treat SSL_write() errors,
https://project.nordu.net/browse/RADSECPROXY-46.